[CONTROLLER/CLOUD] fixes huawei dhcp multi ips bug
