+++ server/controller/cloud/huawei/redis_instance.go
@@ -100,11 +100,18 @@ func (h *HuaWei) getRedisInstances() ([]model.RedisInstance, []model.VInterface,
 				}
 				vifs = append(vifs, vif)
 
+				var subnetLcuuid string
+				for _, subnet := range h.toolDataSet.networkLcuuidToSubnets[networkLcuuid] {
+					if cloudcommon.IsIPInCIDR(ip, subnet.CIDR) {
+						subnetLcuuid = subnet.Lcuuid
+						break
+					}
+				}
 				ip := model.IP{
 					Lcuuid:           common.GenerateUUID(vif.Lcuuid + ip),
 					VInterfaceLcuuid: vif.Lcuuid,
 					IP:               strings.Trim(ip, " "),
-					SubnetLcuuid:     h.toolDataSet.networkLcuuidToSubnetLcuuid[networkLcuuid],
+					SubnetLcuuid:     subnetLcuuid,
 					RegionLcuuid:     regionLcuuid,
 				}
 				ips = append(ips, ip)
